feat(quest): implement separate initialization for main and random quests
- Modify QuestProgressInitializer to handle both main and random quests
- Update PlayerCharacterCreator to use new initialization methods
- Introduce INITIAL_MAIN_QUEST_ID and INITIAL_RAND_QUEST_ID constants
- Separate quest IDs: 1-10 for main quests, 11+ for random quests

This change allows for distinct management of main storyline quests
and randomly generated side quests, improving quest progression tracking.
