+++ app/src/main/java/com/github/cstrerath/uncover/domain/quest/mainquest/QuestProgressInitializer.kt
@@ -7,16 +7,26 @@ import com.github.cstrerath.uncover.data.database.entities.QuestStage
 class QuestProgressInitializer(
     private val characterQuestProgressDao: CharacterQuestProgressDao
 ) {
-    suspend fun initializeQuestProgress(characterId: String) {
-        val initialQuest = CharacterQuestProgress(
+    suspend fun initializeMainQuestProgress(characterId: String) {
+        val initialMainQuest = CharacterQuestProgress(
             characterId = characterId,
-            questId = INITIAL_QUEST_ID,
+            questId = INITIAL_MAIN_QUEST_ID,
             stage = QuestStage.NOT_STARTED
         )
-        characterQuestProgressDao.insertProgress(initialQuest)
+        characterQuestProgressDao.insertProgress(initialMainQuest)
+    }
+
+    suspend fun initializeRandQuestProgress(characterId: String) {
+        val initialRandQuest = CharacterQuestProgress(
+            characterId = characterId,
+            questId = INITIAL_RAND_QUEST_ID,
+            stage = QuestStage.NOT_STARTED
+        )
+        characterQuestProgressDao.insertProgress(initialRandQuest)
     }
 
     companion object {
-        private const val INITIAL_QUEST_ID = 1
+        private const val INITIAL_MAIN_QUEST_ID = 1
+        private const val INITIAL_RAND_QUEST_ID = 11
     }
 }
