+++ app/src/main/java/com/github/cstrerath/uncover/ui/screens/SplashScreen.kt
@@ -1,4 +1,4 @@
-package com.github.cstrerath.uncover.ui.components
+package com.github.cstrerath.uncover.ui.screens
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.LinearProgressIndicator
